Fix a problem that 0th item is selected at the init of Fcitx5
The 0th item in each group is usually set to a simple keyboard such as
`keyboard-us`, but most virtual keyboards using IME add-ons don't use
the 0th item and use only the 1th item, which is the IME for the
keyboard's language.

Such keyboards do not normally select the 0th item, but it is possible
that the 0th item is selected at the initiation of Fcitx5, so this
process is necessary.

This problem occurs when such a keyboard is set for the 0th group.
Example:
```
[Groups/0]
Name=CH
Default Layout=us
DefaultIM=pinyin

[Groups/0/Items/0]
Name=keyboard-us
Layout=

[Groups/0/Items/1]
Name=pinyin
Layout=

[GroupOrder]
0=CH
```

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>